# openapi.yaml
openapi: 3.0.3
info:
  title: Simple Auth Service
  description: |-
    Simple Authentification Service  
    - [The Simple Authentification Service repository](https://github.com/ateachment/to-do)
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 0.0.3
externalDocs:
  description: https://mo6273.schule.hessen.de
  url: https://mo6273.schule.hessen.de
servers:
  [
    {
      url: 'https://localhost:5000/auth/',
      description: 'Local server'
    }
  ]
tags:
  - name: user
    description: Operations about user
  - name: roles
    description: Operations about roles (autorization)
paths:
  /user/login:
    post:
      tags:
        - user
      summary: Logs user into the system
      description: ''
      operationId: loginUser
      requestBody:
        description: Logs user into the system
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/User'
        required: true
      responses:
        '200':
          description: Login sucessful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidToken'
            application/x-www-form-urlencoded:
              schema:
                $ref: '#/components/schemas/ValidToken'
        '403':
          description: Invalid username/password supplied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvalidToken'
  
  
  /user/{token}:
    delete:
      tags:
        - user
      summary: Logs out current logged in user
      description: ''
      operationId: logoutUser
      parameters: 
      - name: token
        in: path
        description: Token that has to be deleted
        required: true
        schema:
          $ref: '#/components/schemas/Token'
      responses:
        '200':
          description: Logout sucessful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvalidToken'



components:
  schemas:
    Token:
      type: string
      example: eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCJ9.eyJleHAiOjE2NzA0NDE4MDMsInJvbGVJRHMiOlsxLDJdfQ.967Mucl7Xrlm_bf33Sh0sq0EzRToZkzsWy5QdvDQU0_JCIkgf1TuC0muKG-ThxihJ-ewMkSvh3u94yDs-RWDMCVavPfLbmdLRJCu6lChI8gVcdR__U61zRBP6cszSFPFK8pJTqs7oe_t6z-HzDMEyxNBwF4bmPuEqXySVETg3lqn8QOb5DNa7056EPWvjUfuKUsPIKfvr_LESMzFXZrnSyNtzQPrSeFYTeynJdl-q1_l9JlqOTiQvqmWbUjyoj8swdJ7brKLuVUzL3yXLV7vBRWhUDLUHzyYlySnZxPrDS_oUpUF8MDOJ9z-WV5d-utDocPmMhMrwP05vXqTYmBGdg
    User:
      type: object
      properties:
        username:
          type: string
          example: testUser
        password:
          type: string
          example: testPwd
    ValidToken:
      type: object
      properties:
        token:
          type: string
          example: eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCJ9.eyJleHAiOjE2NzA0NDE4MDMsInJvbGVJRHMiOlsxLDJdfQ.967Mucl7Xrlm_bf33Sh0sq0EzRToZkzsWy5QdvDQU0_JCIkgf1TuC0muKG-ThxihJ-ewMkSvh3u94yDs-RWDMCVavPfLbmdLRJCu6lChI8gVcdR__U61zRBP6cszSFPFK8pJTqs7oe_t6z-HzDMEyxNBwF4bmPuEqXySVETg3lqn8QOb5DNa7056EPWvjUfuKUsPIKfvr_LESMzFXZrnSyNtzQPrSeFYTeynJdl-q1_l9JlqOTiQvqmWbUjyoj8swdJ7brKLuVUzL3yXLV7vBRWhUDLUHzyYlySnZxPrDS_oUpUF8MDOJ9z-WV5d-utDocPmMhMrwP05vXqTYmBGdg
    InvalidToken:
      type: object
      properties:
        token:
          type: string
          example: -1
    ValidRoleIDs:
      type: object
      properties:
        roleIDs:
          type: string
          example: [1, 2]
    InvalidRoleIDs:
      type: object
      properties:
        roleIDs:
          type: string
          example: -1
  

  requestBodies:
    UserArray:
      description: List of user object
      content:
        application/json:
          schema:
            type: array
            items:
              $ref: '#/components/schemas/User'
              
    
    
    